refactor(chat): move session policy
