+++ packages/chat/src/index.ts
@@ -8,12 +8,16 @@ export {
   normalizeLogEntry,
   normalizeLoggedMessage,
 } from "./log-store.js";
+export { inferConversationKind, resolveChatSessionKey } from "./session-policy.js";
+export type { ResolveSessionKeyOptions } from "./session-policy.js";
 export type {
   AgentMessageContext,
   Attachment,
   BotResponseLogInput,
   ConversationEventRecord,
+  ConversationKind,
   ConversationLogEventType,
+  ConversationLogMessage,
   ConversationLogRecord,
   ConversationLogStoreOptions,
   ConversationTranscriptMessage,
